Rework of methods to compare against
- placed torch implementations on GPU
- removed neural network layers for better comparison to other methods
- Now only comparing matrix multiplication

Note: Everything runs on the GPU now but the methods being compared arent necessarily all sparse/dense as stated in the name. Need to look into the scipy sparse functions a little closer. Also, some are being converted to sparse inside the timer while others aren't.

Matrix size note: pytorch only seems to start making use of GPU memory when the matrix becomes quite large. I'm assuming this is because it's using cache very efficiently and the implemented memory tracking doesn't see that? Or it might just be using the CPU for smaller calculations. We need to look into what we're comparing against a little closer.